backend/requirements.txt
langchain
fastapi[all]
langchain-openai
python-dotenv
youtube-transcript-api